+++ spec/lib/danger/danger_core/environment_manager_spec.rb
@@ -32,8 +32,20 @@ RSpec.describe Danger::EnvironmentManager, use: :ci_helper do
       end
     end
 
-    it "loads Jenkins" do
+    it "loads Jenkins (Github)" do
       with_jenkins_setup_github_and_is_a_pull_request do |system_env|
+        expect(described_class.local_ci_source(system_env)).to eq Danger::Jenkins
+      end
+    end
+
+    it "loads Jenkins (Gitlab)" do
+      with_jenkins_setup_gitlab_and_is_a_pull_request do |system_env|
+        expect(described_class.local_ci_source(system_env)).to eq Danger::Jenkins
+      end
+    end
+
+    it "loads Jenkins (Gitlab v3)" do
+      with_jenkins_setup_gitlab_v3_and_is_a_pull_request do |system_env|
         expect(described_class.local_ci_source(system_env)).to eq Danger::Jenkins
       end
     end
@@ -112,8 +124,20 @@ RSpec.describe Danger::EnvironmentManager, use: :ci_helper do
       end
     end
 
-    it "loads Jenkins" do
+    it "loads Jenkins (Github)" do
       with_jenkins_setup_github_and_is_a_pull_request do |system_env|
+        expect(described_class.pr?(system_env)).to eq(true)
+      end
+    end
+
+    it "loads Jenkins (Gitlab)" do
+      with_jenkins_setup_gitlab_and_is_a_pull_request do |system_env|
+        expect(described_class.pr?(system_env)).to eq(true)
+      end
+    end
+
+    it "loads Jenkins (Gitlab v3)" do
+      with_jenkins_setup_gitlab_v3_and_is_a_pull_request do |system_env|
         expect(described_class.pr?(system_env)).to eq(true)
       end
     end