+++ spec/lib/danger/danger_core/executor_spec.rb
@@ -46,6 +46,12 @@ RSpec.describe Danger::Executor, use: :ci_helper do
         end
       end
 
+      it "not raises error on Jenkins (GitLab v3)" do
+        with_jenkins_setup_gitlab_v3_and_is_a_pull_request do |system_env|
+          expect { described_class.new(system_env).validate!(testing_ui) }.not_to raise_error
+        end
+      end
+
       it "not raises error on Local Git Repo" do
         with_localgitrepo_setup do |system_env|
           ui = testing_ui