+++ spec/support/ci_helper.rb
@@ -70,6 +70,15 @@ module Danger
       def with_jenkins_setup_gitlab_and_is_a_pull_request
         system_env = {
           "JENKINS_URL" => "https://ci.swift.org/job/oss-swift-incremental-RA-osx/lastBuild/",
+          "gitlabMergeRequestIid" => "42"
+        }
+
+        yield(system_env)
+      end
+
+      def with_jenkins_setup_gitlab_v3_and_is_a_pull_request
+        system_env = {
+          "JENKINS_URL" => "https://ci.swift.org/job/oss-swift-incremental-RA-osx/lastBuild/",
           "gitlabMergeRequestId" => "42"
         }